+++ src/common-names.js
@@ -1,3 +1,0 @@
-module.exports = function(target, source = {}) {
-    return (target in source) ? source[target] : null;
-};