+++ src/commonNames.js
@@ -1,0 +1,3 @@
+module.exports = function(target, source = {}) {
+  return target in source ? source[target] : null
+}