+++ src/constants.js
@@ -1,7 +1,7 @@
-const TYPE_IMPORT = 0;
-const TYPE_REQUIRE = 1;
+const TYPE_IMPORT = 'import'
+const TYPE_REQUIRE = 'require'
 
 module.exports = {
-    TYPE_IMPORT,
-    TYPE_REQUIRE,
-};
+  TYPE_IMPORT,
+  TYPE_REQUIRE
+}