+++ src/is-require.js
@@ -1,3 +1,0 @@
-module.exports = function isRequire(line) {
-    return line.match(/require\(/) || line.match(/^import/);
-};