+++ src/isRequire.js
@@ -1,0 +1,3 @@
+module.exports = function isRequire(line) {
+  return line.match(/require\(/) || line.match(/^import/)
+}