+++ test/testFiles/fileWithBlockComment.js
@@ -1,0 +1,6 @@
+/*
+i am
+a
+block comment
+*/
+const a = "var";