+++ test/testFiles/fileWithDestructuringImport.js
@@ -1,0 +1,10 @@
+import {
+  forEach,
+  get,
+  set,
+  omit,
+  filter,
+  filterBy,
+  concat,
+  difference
+} from "./test";