+++ test/testFiles/fileWithImports.js
@@ -1,0 +1,2 @@
+import Promise from "bluebird";
+import something from "./fileWithBlockComment";