+++ CHANGELOG.md
@@ -1,3 +1,52 @@
+# 1.2.10
+
+- FIXED: places external import before local destructured imports.
+
+# 1.2.8
+
+- ADDED: option to insert the require statement at cursor (insertAtCursor).
+
+# 1.2.4
+
+- ADDED: destructuring import now goes deeper. Will get named exports from files exported
+  from selected file.
+
+# 1.2.0
+
+- ADDED: destructuring import command to require named exports from files.
+
+# 1.1.3
+
+- FIXED: add a newline before if insert position is at end of file.
+
+# 1.1.2
+
+- ADDED: quote type config and quote type autodetection.
+- ADDED: semicolon autodetection.
+
+# 1.1.0
+
+- ADDED: ability to require multiple modules/files with a single command.
+
+# 1.8.0
+
+- ADDED: support for .ts/.tsx file types.
+
+# 1.0.5
+
+- ADDED: config option for semicolon.
+- FIXED: position when a block comment is at the top of a file.
+
+# 1.0.0
+
+- forked repository.
+- FIXED: position when a named import is present in the file.
+
+# 0.7.0
+
+- Absolute imports are inserted before relative imports (by **@nickyhajal**).
+- Allow aliases for local imports (by **@nickyhajal**).
+
 # 0.6.2
 
 - FIX: Skip 'use strict' before inserting the require (thx to **@jhnns**).